pyproject.toml
[project]
name = "personal-website"
version = "0.1.0"
description = "Personal website of Kyle Wong"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit>=1.32.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pillow>=10.0.0",
]

[tool.uv]
dev-dependencies = []
